remove unnecessary checks
Since pest will only return something that is either matching the given grammar
or an error, we do not have to check the rules again ourselves.
Since the functions are still public, added some asserts to ensure this.
These functions should be made private and the asserts could be removed (or
turned into debug_assert's for debugging purposes).
resolves #3